# pgauthzd/internal/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: pgauthzd
  description: >
    PostgreSQL-native Zanzibar-style authorization over HTTP: the pgauthz
    native API (raw graph decisions, search, explain, watch, tuple writes)
    and the AuthZEN 1.0 access evaluation/search API.
  # Placeholder — replaced at startup with the build version (the same value
  # build_info reports); see api.SetOpenAPIVersion.
  version: 0.0.0-dev
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0

# Default: the public listener's end-user/service JWT. Native operations
# additionally accept the shared service token — that is how the SAME paths
# are authenticated on the internal callback listener (declared per
# operation below as bearerAuth OR serviceToken). Discovery, health, and this
# document are public (security: []).
security:
  - bearerAuth: []
...
paths:
...
  /access/v1/evaluation:
    post:
      tags: [authzen]
      operationId: evaluation
      summary: Single access evaluation (AuthZEN §5)
      parameters:
        - $ref: '#/components/parameters/StoreHeader'
        - $ref: '#/components/parameters/DetailHeader'
        - $ref: '#/components/parameters/ConsistencyHeader'
        - $ref: '#/components/parameters/RevisionHeader'
      requestBody: { $ref: '#/components/requestBodies/Evaluation' }
      responses:
        '200': { $ref: '#/components/responses/EvaluationOK' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/FreshnessConflict' }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /pgauthz/v1/explain:
    post:
      tags: [native-read]
      security:
        - bearerAuth: []
        - serviceToken: []
      x-pgauthz-availability:
        listeners: [public, internal-callback]
        public_listener: not-when-opa-fronted
      operationId: explain
      summary: WHY was access allowed or denied (decision + trace)
      parameters:
        - $ref: '#/components/parameters/StoreHeader'
        - $ref: '#/components/parameters/ConsistencyHeader'
        - $ref: '#/components/parameters/RevisionHeader'
      requestBody: { $ref: '#/components/requestBodies/Explain' }
      responses:
        '200': { $ref: '#/components/responses/ExplainOK' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/FreshnessConflict' }
        '500': { $ref: '#/components/responses/ServerError' }
        '501': { $ref: '#/components/responses/NotImplemented' }
...
  /pgauthz/v1/write:
    post:
      tags: [native-write]
      security:
        - bearerAuth: []
        - serviceToken: []
      x-pgauthz-availability:
        listeners: [public, internal-callback]
        public_listener: not-when-opa-fronted
        profile: full
      operationId: writeTuples
      summary: Batch-upsert relationship tuples
      parameters:
        - $ref: '#/components/parameters/StoreHeader'
      requestBody: { $ref: '#/components/requestBodies/WriteTuples' }
      responses:
        '200': { $ref: '#/components/responses/WriteOK' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
        '501': { $ref: '#/components/responses/NotImplemented' }
...
  /pgauthz/v1/delete:
    post:
      tags: [native-write]
      security:
        - bearerAuth: []
        - serviceToken: []
      x-pgauthz-availability:
        listeners: [public, internal-callback]
        public_listener: not-when-opa-fronted
        profile: full
      operationId: deleteTuples
      summary: Batch-delete relationship tuples
      parameters:
        - $ref: '#/components/parameters/StoreHeader'
      requestBody: { $ref: '#/components/requestBodies/WriteTuples' }
      responses:
        '200': { $ref: '#/components/responses/DeleteOK' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
        '501': { $ref: '#/components/responses/NotImplemented' }
...
  /pgauthz/v1/delete-user:
    post:
      tags: [native-write]
      security:
        - bearerAuth: []
        - serviceToken: []
      x-pgauthz-availability:
        listeners: [public, internal-callback]
        public_listener: not-when-opa-fronted
        profile: full
      operationId: deleteUserTuples
      summary: Offboarding — remove every tuple for a subject
      parameters:
        - $ref: '#/components/parameters/StoreHeader'
      requestBody: { $ref: '#/components/requestBodies/DeleteUser' }
      responses:
        '200': { $ref: '#/components/responses/DeleteOK' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
        '501': { $ref: '#/components/responses/NotImplemented' }
...
  /pgauthz/v1/write-checked:
    post:
      tags: [native-write]
      security:
        - bearerAuth: []
        - serviceToken: []
      x-pgauthz-availability:
        listeners: [public, internal-callback]
        public_listener: not-when-opa-fronted
        profile: full
      operationId: writeTuplesChecked
      summary: Conditional/atomic write — preconditions gate deletes+writes
      parameters:
        - $ref: '#/components/parameters/StoreHeader'
      requestBody: { $ref: '#/components/requestBodies/WriteChecked' }
      responses:
        '200': { $ref: '#/components/responses/WriteCheckedOK' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
        '501': { $ref: '#/components/responses/NotImplemented' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        PUBLIC listener: an end-user/service JWT from a trusted issuer
        (multi-issuer via JWT_ISSUERS). Per-issuer store and db-role bindings
        apply; writes additionally require the WRITER_ROLE claim.
    serviceToken:
      type: http
      scheme: bearer
      description: >
        INTERNAL callback listener only — sent to the PUBLIC listener it is
        rejected (401, it is not a JWT). The shared service credential
        (Authorization: Bearer <INTERNAL_SERVICE_TOKEN>), optionally layered
        under mTLS. The listener trusts its OPA upstream — it does NOT
        re-verify the end-user JWT, and honors the X-PGAuthz-Role header
        (per-app DB role) asserted by OPA. Never expose this listener to
        untrusted callers.
  parameters:
...
    StoreHeader:
      name: X-PGAuthz-Store
      in: header
      required: false
      schema: { type: string }
      description: >
        Store selection for unscoped paths: the /stores/{store} path segment
        wins, then this header, then the instance's DEFAULT_STORE.
    ConsistencyHeader:
      name: X-PGAuthz-Consistency
      in: header
      required: false
      schema:
        type: string
        enum: [minimize_latency, at_least_as_fresh]
      description: >
        Read consistency. `at_least_as_fresh` REQUIRES an
        X-PGAuthz-Revision token and demands read-your-writes: the node serves
        the read only if it has replayed to the token's WAL position on the
        token's timeline; otherwise 409 (or transparent primary fallback).
        Default: minimize_latency (answer locally).
    RevisionHeader:
      name: X-PGAuthz-Revision
      in: header
      required: false
      schema: { type: string }
      description: >
        The opaque freshness token returned by an earlier write; required with
        `X-PGAuthz-Consistency: at_least_as_fresh`.
    DetailHeader:
      name: X-PGAuthz-Detail
      in: header
      required: false
      schema: { type: string, enum: ["true"] }
      description: >
        Opt into the rich decision detail on the AuthZEN evaluation response
        (state allow|deny|conditional, missing_context, conditions, reason,
        model) — returned in the response `context` field.
  headers:
    XPGAuthzRevision:
      schema: { type: string }
      description: >
        Freshness token minted for this write (HMAC-signed {kid, timeline,
        LSN}). Present a later read with it via X-PGAuthz-Revision +
        X-PGAuthz-Consistency: at_least_as_fresh for read-your-writes.
    XPGAuthzRevisionStatus:
      schema:
        type: string
        enum: [issued, unavailable, disabled]
      description: >
        Mint outcome for this write. `issued` = token in X-PGAuthz-Revision;
        `unavailable` = freshness is enabled but minting failed (the write
        itself committed); `disabled` = freshness not configured.
    XPGAuthzStale:
      schema:
        type: string
        enum: [stale, wrong_epoch, unknown]
      description: The freshness verdict that caused a 409.
    XPGAuthzServedBy:
      schema: { type: string, enum: [primary] }
      description: >
        Present when the read was transparently served from the primary
        fallback pool because the local replica wasn't fresh enough.
  requestBodies:
    Evaluation:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EvaluationRequest' }
...
    Explain:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ExplainRequest' }
...
    WriteTuples:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/WriteTuplesRequest' }
    DeleteUser:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DeleteUserRequest' }
    WriteChecked:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/WriteCheckedRequest' }
  responses:
    EvaluationOK:
      description: The access decision.
      headers:
        X-PGAuthz-Served-By: { $ref: '#/components/headers/XPGAuthzServedBy' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EvaluationResponse' }
...
    ExplainOK:
      description: >
        The engine's explain_access JSON verbatim: result, decision
        {allowed, reason}, summary (human-readable trace), tree, steps.
      headers:
        X-PGAuthz-Served-By: { $ref: '#/components/headers/XPGAuthzServedBy' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ExplainResponse' }
...
    WriteOK:
      description: Tuples written.
      headers:
        X-PGAuthz-Revision: { $ref: '#/components/headers/XPGAuthzRevision' }
        X-PGAuthz-Revision-Status: { $ref: '#/components/headers/XPGAuthzRevisionStatus' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/WriteResponse' }
    DeleteOK:
      description: Tuples deleted.
      headers:
        X-PGAuthz-Revision: { $ref: '#/components/headers/XPGAuthzRevision' }
        X-PGAuthz-Revision-Status: { $ref: '#/components/headers/XPGAuthzRevisionStatus' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/DeleteResponse' }
    WriteCheckedOK:
      description: >
        The engine's write_tuples_checked JSON verbatim — applied (bool),
        precondition results, and counts. The freshness token rides the
        X-PGAuthz-Revision header only.
      headers:
        X-PGAuthz-Revision: { $ref: '#/components/headers/XPGAuthzRevision' }
        X-PGAuthz-Revision-Status: { $ref: '#/components/headers/XPGAuthzRevisionStatus' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/WriteCheckedResponse' }
    BadRequest:
      description: >
        Malformed request — invalid JSON, missing required fields, an invalid
        consistency mode, or an invalid/forged/retired freshness token (one
        opaque message for every bad-token cause; the reason is logged
        server-side only).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing/invalid bearer token (or service token on the internal listener).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: >
        Authorization failure — issuer store/db-role binding violation,
        missing WRITER_ROLE claim on a public write, SEARCH_REQUIRED_ROLE not
        held, a forbidden per-app DB role, or a write to a read-only
        (decision-only) instance.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    FreshnessConflict:
      description: >
        An at_least_as_fresh read this node cannot satisfy. The
        verdict picks the recovery: stale/unknown → retry the primary or
        another replica; wrong_epoch → NOT retryable, and NOT fixed by an
        unrelated new token: a failover happened since the write and the
        write itself may have been lost — re-read and reconcile the intended
        authorization state on the current primary (reapply the change
        idempotently if missing), then use that write's token. This matters
        most for revokes/offboarding. primary_consulted=true means the
        transparent fallback already re-checked the primary and it cannot
        satisfy the token either.
      headers:
        X-PGAuthz-Stale: { $ref: '#/components/headers/XPGAuthzStale' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/FreshnessConflict' }
    ServerError:
      description: Internal error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotImplemented:
      description: >
        The operation requires a capability this instance does not have
        (e.g. native reads without a direct pgx backend, or native writes
        without the full profile's writer backend).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    Subject:
      type: object
      properties:
        type: { type: string, description: Subject type (e.g. internal_user). }
        id: { type: string, description: Subject identifier. }
        properties:
          type: object
          additionalProperties: true
          description: >
            AuthZEN-standard extension attributes. Accepted for compatibility;
            pgauthz does not evaluate them — pass condition input via the
            request-level context instead.
      description: >
        On the public listener a request subject must match the JWT-derived
        subject unless ALLOW_SUBJECT_OVERRIDE is enabled (trusted-PEP mode);
        an omitted subject defaults to the authenticated one (the official
        AuthZEN schema requires subject — pgauthzd is deliberately more
        lenient here).
    Action:
      type: object
      properties:
        name: { type: string, description: The relation/permission to test (e.g. can_read). }
        properties:
          type: object
          additionalProperties: true
          description: Accepted for AuthZEN compatibility; not evaluated.
    Resource:
      type: object
      properties:
        type: { type: string }
        id: { type: string }
        properties:
          type: object
          additionalProperties: true
          description: Accepted for AuthZEN compatibility; not evaluated.
    Context:
      type: object
      additionalProperties: true
      description: Request context for conditional (ABAC) tuples.
...
    EvaluationRequest:
      type: object
      properties:
        subject: { $ref: '#/components/schemas/Subject' }
        action: { $ref: '#/components/schemas/Action' }
        resource: { $ref: '#/components/schemas/Resource' }
        context: { $ref: '#/components/schemas/Context' }
    EvaluationResponse:
      type: object
      properties:
        decision: { type: boolean }
        context:
          type: object
          additionalProperties: true
          description: >
            Rich decision detail — a pgauthz EXTENSION returned only when the
            caller opts in via X-PGAuthz-Detail: state allow|deny|conditional,
            missing_context, conditions, reason, model. Note this is not the
            official AuthZEN response-context reason envelope ({id,
            reason_admin, reason_user}); callers that never send the header
            never see a context field.
      required: [decision]
...
    ExplainRequest:
      type: object
      properties:
        subject: { $ref: '#/components/schemas/Subject' }
        action: { $ref: '#/components/schemas/Action' }
        resource: { $ref: '#/components/schemas/Resource' }
        context: { $ref: '#/components/schemas/Context' }
      required: [action, resource]
    ExplainResponse:
      type: object
      additionalProperties: true
      description: >
        authz.explain_access JSON verbatim — result (bool), decision
        {allowed, reason}, summary (human-readable), tree (recursive trace),
        steps (flat step list).
...
    TupleInput:
      type: object
      properties:
        user_type: { type: string }
        user_id: { type: string }
        user_relation:
          type: string
          description: Set for userset subjects (e.g. team#member → 'member').
        relation: { type: string }
        object_type: { type: string }
        object_id: { type: string }
        condition: { type: string, description: Name of a condition gating this tuple. }
        context: { $ref: '#/components/schemas/Context' }
        expires_at: { type: string, format: date-time }
      required: [user_type, user_id, relation, object_type, object_id]
    PerformedBy:
      type: string
      description: >
        Audit author (the immutable trail's performed_by). PUBLIC listener:
        the authenticated JWT subject is authoritative — omit this field or
        send the same value; a DIFFERING value is rejected (403) unless
        ALLOW_SUBJECT_OVERRIDE is enabled (trusted-PEP deployments asserting
        on-behalf-of actors). INTERNAL callback listener: the trusted OPA
        upstream passes the subject it authenticated here.
    Consistency:
      type: string
      enum: [applied, durable, eventual]
      description: >
        Per-transaction durability (maps to synchronous_commit): applied =
        remote_apply (strict revocation), durable = on, eventual = local.
        Omitted = the connection default.
    WriteTuplesRequest:
      type: object
      properties:
        tuples:
          type: array
          items: { $ref: '#/components/schemas/TupleInput' }
          minItems: 1
        consistency: { $ref: '#/components/schemas/Consistency' }
        performed_by: { $ref: '#/components/schemas/PerformedBy' }
      required: [tuples]
    WriteResponse:
      type: object
      properties:
        store: { type: string }
        written: { type: integer }
        revision:
          type: string
          description: Freshness token for this write (when enabled).
      required: [store, written]
    DeleteResponse:
      type: object
      properties:
        store: { type: string }
        deleted: { type: integer }
        revision: { type: string }
      required: [store, deleted]
    DeleteUserRequest:
      type: object
      properties:
        user: { $ref: '#/components/schemas/Subject' }
        consistency: { $ref: '#/components/schemas/Consistency' }
        performed_by: { $ref: '#/components/schemas/PerformedBy' }
      required: [user]
    WriteCheckedRequest:
      type: object
      properties:
        preconditions:
          type: array
          items:
            type: object
            additionalProperties: true
          description: >
            Tuple-existence preconditions (write_tuples_checked shape); all
            must hold or nothing is applied.
        deletes:
          type: array
          items: { $ref: '#/components/schemas/TupleInput' }
        writes:
          type: array
          items: { $ref: '#/components/schemas/TupleInput' }
        consistency: { $ref: '#/components/schemas/Consistency' }
        performed_by: { $ref: '#/components/schemas/PerformedBy' }
    WriteCheckedResponse:
      type: object
      additionalProperties: true
      description: >
        authz.write_tuples_checked JSON verbatim — applied (bool), failed
        precondition detail, written/deleted counts.

    # ── errors & meta ────────────────────────────────────────────────────────
    Error:
      type: object
      properties:
        status: { type: integer }
        message: { type: string }
      required: [status, message]
    FreshnessConflict:
      type: object
      properties:
        status: { type: integer }
        error:
          type: string
          enum: [freshness_constraint_unsatisfied]
        verdict:
          type: string
          enum: [stale, wrong_epoch, unknown]
        primary_consulted:
          type: boolean
          description: The transparent fallback already re-checked the primary.
        message: { type: string, description: Per-verdict recovery guidance. }
      required: [status, error, verdict, primary_consulted, message]
...